console.log instead of console.warn
The `console.warn()` function is an alias for `console.error()`. https://nodejs.org/api/console.html#console_console_warn_data and `console.error` prints to `stderr`. This is a massive pain in automated build environments which fails the build when there is something written to `stderr`. More than anything, it's not logical to warn for all these cases as no of them are warning. They are info.