fix: Messages for success are printed on stderr
In case the clean operation is successful, the messages for success are printed on stderr. This is misconception as all success messages should be printed on stdout.
Many tools check the stderr output and print errors or even break the process in case there's something printed there.
The current PR changes the method used to print success messages - instead of using console.warn, use console.log.
This will allow external tools to use the stderr output as indication that there was a problem. When the only output is on stdout, they can consider the operation as successful.